Make stack the primary download method on the Downloads page
As it stands, the current downloads page contradicts much online advice
about getting set up with Haskell, leading to user confusion. More
motivation to be brought up in the pull request.

Currently links to pages on github.com for both the stack guide and
download instructions. These are Markdown pages that could easily be
copied into the hl codebase instead to keep all links local. I can make
that change if desired, but didn't want to bloat this initial work with
a bunch of unrelevant line changes.